+++ tests/origocli/commands/status_test.py
@@ -1,8 +1,8 @@
-import pytest
+from copy import deepcopy
+
 from unittest.mock import ANY
 from conftest import set_argv
 from okdata.cli.commands.status import StatusCommand
-from okdata.sdk.status import Status
 from okdata.cli.output import TableOutput
 
 
@@ -100,6 +100,15 @@ failed_trace = [
         "trace_event_status": "FAILED",
         "component": "pipeline-instance",
         "trace_status": "FINISHED",
+        "errors": [
+            {
+                "message": {
+                    "nb": "Ugyldig verdi.",
+                    "nn": "Verdi ikkje gangbar.",
+                    "en": "Invalid value.",
+                }
+            }
+        ],
         "user": "system",
         "domain_id": "na",
         "start_time": "2020-01-01T11:45:13.939324",
@@ -111,20 +120,20 @@ failed_trace = [
 ]
 
 
-def test_get_status(mock_status_sdk, mocker, mock_print):
+def test_get_status(mocker, mock_print):
+    trace = deepcopy(successful_trace)
     set_argv("status", successful_trace_id)
     cmd = StatusCommand()
-    mocker.spy(cmd.sdk, "get_status")
+    mocker.patch.object(cmd, "sdk")
+    cmd.sdk.get_status.return_value = trace
     mocker.spy(cmd, "latest_event_for_status")
     mocker.spy(StatusCommand, "find_latest_event")
     mocker.spy(TableOutput, "add_row")
     cmd.handler()
     cmd.sdk.get_status.assert_called_once_with(successful_trace_id)
-    cmd.latest_event_for_status.assert_called_once_with(
-        successful_trace_id, successful_trace
-    )
-    StatusCommand.find_latest_event.assert_called_once_with(successful_trace)
-    expected_event = successful_trace[-1]
+    cmd.latest_event_for_status.assert_called_once_with(successful_trace_id, trace)
+    StatusCommand.find_latest_event.assert_called_once_with(trace)
+    expected_event = trace[-1]
     TableOutput.add_row.assert_called_once_with(
         ANY,
         {
@@ -132,24 +141,25 @@ def test_get_status(mock_status_sdk, mocker, mock_print):
             "trace_id": expected_event["trace_id"],
             "trace_status": expected_event["trace_status"],
             "trace_event_status": expected_event["trace_event_status"],
+            "errors": [],
         },
     )
 
 
-def test_get_status_in_progress(mock_status_sdk, mocker, mock_print):
+def test_get_status_in_progress(mocker, mock_print):
+    trace = deepcopy(in_progress_trace)
     set_argv("status", in_progress_trace_id)
     cmd = StatusCommand()
-    mocker.spy(cmd.sdk, "get_status")
+    mocker.patch.object(cmd, "sdk")
+    cmd.sdk.get_status.return_value = trace
     mocker.spy(cmd, "latest_event_for_status")
     mocker.spy(StatusCommand, "find_latest_event")
     mocker.spy(TableOutput, "add_row")
     cmd.handler()
     cmd.sdk.get_status.assert_called_once_with(in_progress_trace_id)
-    cmd.latest_event_for_status.assert_called_once_with(
-        in_progress_trace_id, in_progress_trace
-    )
-    StatusCommand.find_latest_event.assert_called_once_with(in_progress_trace)
-    expected_event = in_progress_trace[-1]
+    cmd.latest_event_for_status.assert_called_once_with(in_progress_trace_id, trace)
+    StatusCommand.find_latest_event.assert_called_once_with(trace)
+    expected_event = trace[-1]
     TableOutput.add_row.assert_called_once_with(
         ANY,
         {
@@ -157,22 +167,25 @@ def test_get_status_in_progress(mock_status_sdk, mocker, mock_print):
             "trace_id": expected_event["trace_id"],
             "trace_status": expected_event["trace_status"],
             "trace_event_status": expected_event["trace_event_status"],
+            "errors": [],
         },
     )
 
 
-def test_get_status_failed(mock_status_sdk, mocker):
+def test_get_status_failed(mocker):
+    trace = deepcopy(failed_trace)
     set_argv("status", failed_trace_id)
     cmd = StatusCommand()
-    mocker.spy(cmd.sdk, "get_status")
+    mocker.patch.object(cmd, "sdk")
+    cmd.sdk.get_status.return_value = trace
     mocker.spy(cmd, "latest_event_for_status")
     mocker.spy(StatusCommand, "find_latest_event")
     mocker.spy(TableOutput, "add_row")
     cmd.handler()
     cmd.sdk.get_status.assert_called_once_with(failed_trace_id)
-    cmd.latest_event_for_status.assert_called_once_with(failed_trace_id, failed_trace)
-    StatusCommand.find_latest_event.assert_called_once_with(failed_trace)
-    expected_event = failed_trace[-1]
+    cmd.latest_event_for_status.assert_called_once_with(failed_trace_id, trace)
+    StatusCommand.find_latest_event.assert_called_once_with(trace)
+    expected_event = trace[-1]
     TableOutput.add_row.assert_called_once_with(
         ANY,
         {
@@ -180,30 +193,35 @@ def test_get_status_failed(mock_status_sdk, mocker):
             "trace_id": expected_event["trace_id"],
             "trace_status": expected_event["trace_status"],
             "trace_event_status": expected_event["trace_event_status"],
+            "errors": ["Invalid value."],
         },
     )
 
 
-def test_get_status_with_history(mock_status_sdk, mocker):
+def test_get_status_with_history(mocker):
+    trace = deepcopy(successful_trace)
     set_argv("status", successful_trace_id, "--history")
     cmd = StatusCommand()
-    mocker.spy(cmd.sdk, "get_status")
+    mocker.patch.object(cmd, "sdk")
+    cmd.sdk.get_status.return_value = trace
     mocker.spy(cmd, "full_history_for_status")
     mocker.spy(TableOutput, "add_rows")
     cmd.handler()
     cmd.sdk.get_status.assert_called_once_with(successful_trace_id)
     cmd.full_history_for_status.assert_called_once()
-    TableOutput.add_rows.assert_called_once_with(ANY, successful_trace)
+    TableOutput.add_rows.assert_called_once_with(ANY, trace)
 
 
-@pytest.fixture()
-def mock_status_sdk(monkeypatch):
-    def get_status_return(self, uuid):
-        if uuid == failed_trace_id:
-            return failed_trace
-        elif uuid == in_progress_trace_id:
-            return in_progress_trace
-        else:
-            return successful_trace
-
-    monkeypatch.setattr(Status, "get_status", get_status_return)
+def test_get_status_failed_with_history(mocker):
+    trace = deepcopy(failed_trace)
+    set_argv("status", failed_trace_id, "--history")
+    cmd = StatusCommand()
+    mocker.patch.object(cmd, "sdk")
+    cmd.sdk.get_status.return_value = trace
+    mocker.spy(cmd, "full_history_for_status")
+    mocker.spy(TableOutput, "add_rows")
+    cmd.handler()
+    cmd.sdk.get_status.assert_called_once_with(failed_trace_id)
+    cmd.full_history_for_status.assert_called_once()
+    assert trace[-1]["errors"] == ["Invalid value."]
+    TableOutput.add_rows.assert_called_once_with(ANY, trace)